feat(large dataset): use a heuristic approach to limit memory usage for CXG creation. (#4634)
* Remove depreciated tiledb configuration parameters

The tiledb default is the number of cores. This should be sufficient.

* Use WMG tiledb configuration to control memory used for consolidation

- set deduplicate to reduce tiledb requirements when consolidating.

---------
